refactor(analysis): split dimension_runner cache orchestration
