+++ tests/tools/test_logging_boundary.py
@@ -39,6 +39,7 @@ DECLARED_LOGGING_SITES: dict[str, str] = {
     'analysis/_runner_markers.py': 'Structured marker emission and heartbeat callback for the runner pipeline - out of scope for this sweep (not a flagged per-site conversion)',
     'analysis/api_prompt_assembly.py': 'Prompt assembly for the direct API runner - out of scope for this sweep (not a flagged per-site conversion)',
     'analysis/cache/_failure_streak.py': 'Consecutive-failure circuit breaker for the dim runner - out of scope for this sweep (not a flagged per-site conversion)',
+    'analysis/cache/_replay.py': "Cache-replay path: writing cached findings and their events.jsonl mirror back into a run's evidence - split out of dimension_runner.py, inherits its out-of-scope logging (not a flagged per-site conversion)",
     'analysis/cache/cache_writer.py': 'Factory for the per-file cache-write callback passed to FindingsRouter - out of scope for this sweep (not a flagged per-site conversion)',
     'analysis/cache/consolidation.py': "Consolidation state - flip a completed run's cache entries - out of scope for this sweep (not a flagged per-site conversion)",
     'analysis/cache/dimension_helpers.py': 'Dimension-level cache helpers bridging RunConfig and the filesystem - out of scope for this sweep (not a flagged per-site conversion)',
